Adjust multipacket return types to be more appropriate

+++ xbd_multipacket.c
@@ -60,7 +60,7 @@ size_t XBD_genSucessiveMultiPacket(struct xbd_multipkt_state *state, void* dstbu
 	return offset;
 }
 
-size_t XBD_recInitialMultiPacket(struct xbd_multipkt_state *state, const void *recdata, uint32_t reclen, const char *cmd_code, bool hastype, bool hasaddr) {
+int XBD_recInitialMultiPacket(struct xbd_multipkt_state *state, const void *recdata, uint32_t reclen, const char *cmd_code, bool hastype, bool hasaddr) {
 	uint32_t offset=0;
 
 	state->seqn=0;
@@ -96,7 +96,7 @@ size_t XBD_recInitialMultiPacket(struct xbd_multipkt_state *state, const void *r
 	return 0;
 }
 
-size_t XBD_recSucessiveMultiPacket(struct xbd_multipkt_state *state, const void *recdata, uint32_t reclen, void *dstbuf, uint32_t dstlenmax, const char *code) {
+int XBD_recSucessiveMultiPacket(struct xbd_multipkt_state *state, const void *recdata, uint32_t reclen, void *dstbuf, uint32_t dstlenmax, const char *code) {
 	uint32_t offset=0;
 	uint32_t cpylen;
